Fix god trait.
