+++ src/Task/ContextRunTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Wunderio\GrumPHP\Task;
+
+use GrumPHP\Task\Context\ContextInterface;
+use GrumPHP\Task\Context\GitPreCommitContext;
+use GrumPHP\Task\Context\RunContext;
+
+/**
+ * Trait ContextRunTrait.
+ *
+ * @package Wunderio\GrumPHP\Task
+ */
+trait ContextRunTrait {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function canRunInContext(ContextInterface $context): bool {
+    return $context instanceof GitPreCommitContext || $context instanceof RunContext;
+  }
+
+}
