+++ tests/ContextRunTraitTest.php
@@ -1,0 +1,41 @@
+<?php
+
+/**
+ * @file
+ * Tests covering ContextRunTrait.
+ */
+
+declare(strict_types = 1);
+
+use GrumPHP\Collection\FilesCollection;
+use GrumPHP\Task\Context\GitCommitMsgContext;
+use GrumPHP\Task\Context\GitPreCommitContext;
+use GrumPHP\Task\Context\RunContext;
+use PHPUnit\Framework\TestCase;
+use Wunderio\GrumPHP\Task\ContextRunTrait;
+
+/**
+ * Class ContextRunTraitTest.
+ */
+final class ContextRunTraitTest extends TestCase {
+
+  /**
+   * Test run contexts.
+   *
+   * @covers \Wunderio\GrumPHP\Task\ContextRunTrait::canRunInContext
+   */
+  public function testRunsInGitAndRunContexts(): void {
+    $stub = $this->getMockBuilder(ContextRunTrait::class)
+      ->onlyMethods([
+        'canRunInContext',
+      ])
+      ->setMethodsExcept(['canRunInContext'])
+      ->getMockForTrait();
+    $this->assertTrue($stub->canRunInContext(new RunContext(new FilesCollection())));
+    $this->assertTrue($stub->canRunInContext(new GitPreCommitContext(new FilesCollection())));
+
+    $commitMessageContext = $this->createMock(GitCommitMsgContext::class);
+    $this->assertFalse($stub->canRunInContext($commitMessageContext));
+  }
+
+}
